fix the issue where the new tweet post isn't displaying on the top of the list. enhance the UI
