+++ app/src/main/res/layout/activity_timeline.xml
@@ -26,7 +26,7 @@
             android:layout_height="wrap_content"
             android:layout_marginBottom="25dp"
             android:src="@drawable/ic_add_white_24dp"
-            app:backgroundTint="@android:color/holo_blue_dark"
+            app:backgroundTint="@android:color/holo_blue_light"
             app:fabSize="normal"
             app:layout_constraintBottom_toBottomOf="parent"
             android:layout_marginEnd="@dimen/fab_margin"
