+++ app/src/main/res/values/dimens.xml
@@ -7,5 +7,7 @@
     <dimen name="fab_margin">16dp</dimen>
     <dimen name="btn_padding">10dp</dimen>
     <dimen name="btn_margin">10dp</dimen>
+    <dimen name="iv_margin">5dp</dimen>
+    <dimen name="layout_padding">10dp</dimen>
 
 </resources>
