+++ app/src/main/res/values/styles.xml
@@ -10,7 +10,10 @@
 
     <!-- Application theme. -->
     <style name="AppTheme" parent="AppBaseTheme">
-        <!-- All customizations that are NOT specific to a particular API-level can go here. -->
+        <item name="colorPrimary">@android:color/holo_blue_light</item>
+        <item name="colorPrimaryDark">@android:color/holo_blue_dark</item>
+        <item name="colorAccent">@android:color/white</item>
+    <!-- All customizations that are NOT specific to a particular API-level can go here. -->
     </style>
 
 </resources>
